Index.html and style.css
Main layout for mobile first almost completed, will have to test on mobile. Background svg and color is set for both mobile and desktop. I also need to set up the social media links at the bottom of the mobile view. pretty happy with what is started. Fonts and sizing is set for mobile, may have to tweak later.